email validation in front, fixed pull request notes
