+++ app/views/layout.erb
@@ -7,7 +7,7 @@
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV" crossorigin="anonymous"></script>
    
     <script src="/js/jquery_maskedinput.js" type="text/javascript"></script>
-    <script src="/js/file_read.js" type="text/javascript">></script>
+    <script src="/js/add_functions.js" type="text/javascript">></script>
     <!-- CSS only -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
   </head>
@@ -19,5 +19,6 @@
       <% end %>      
       <%= yield %>
     </div>
+    <script src="/js/onload_events.js" type="text/javascript">></script>
   </body>
 </html>
