+++ public/js/add_functions.js
@@ -1,3 +1,6 @@
+ 
+// Images
+
 function handleFiles(files) {    
     var preview = document.getElementById("preview");
 
@@ -11,7 +14,8 @@ function handleFiles(files) {
         if (!file.type.startsWith('image/')){ continue }
             
         var img = document.createElement("img");
-        img.classList.add("obj");
+        img.classList.add("img-fluid");
+        img.style.maxWidth = "200px"
         img.file = file;
 
         preview.appendChild(img); // Предполагается, что "preview" это div, в котором будет отображаться содержимое.
@@ -20,4 +24,22 @@ function handleFiles(files) {
         reader.onload = (function(aImg) { return function(e) { aImg.src = e.target.result; }; })(img);
         reader.readAsDataURL(file);
     }
+}
+
+// Inputs
+
+function ValidateEmail(mail) 
+{     
+    if(mail==""){
+        return (true)
+    }
+    
+    if (/^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/.test(mail))
+    {
+        document.getElementById("email_error").hidden = true;
+        return (true)
+    }
+    
+    document.getElementById("email_error").hidden = false;
+    return (false)
 }
