+++ public/js/onload_events.js
@@ -1,0 +1,6 @@
+
+// On Document load events: input masks
+
+$(function(){
+    $("#candidate_phone").mask("+375 (99) 999-99-99")
+});
